Adding a `bottomOffset` onto react-waypoint
Because we do not have any children inside of the Waypoint, it essentially
acts like a horizontal line across the page. As such when we are right at
the bottom of the page, there is nowhere for the Waypoint to scroll into view.

Adding an offset means that if we're at the bottom of the page, it'll render
the last endpoint.