Fix issue where cursor lost focus on typing
This was due to a problem similar to the following: https://stackoverflow.com/a/49688084

As we are dynamically determining what input types to use at runtime based on the oas
file, this was causing the input elements to be totally recreated on every render, with
the correct value inside, but with the cursor focus lost.

The solution is to create the input types once, one level up in the Doc constructor and
reuse this component on each re-render.

There's no way to get document.activeElement in enzyme, so there's no easy way for me
to test agains this regression.