Fix issue with page refreshing on auth box open
This was really difficult to debug and I couldn't reproduce it but suspected
it was something to do with this <a> tag missing a href.

Greg figured it out yesterday that it was when you start on another page of ReadMe
then navigate to a reference page, this issue occurs.

I'm not 100% sure of the cause of the bug, i'm guessing it's something to do with
the pjax library highjacking all clicks on anchors, and because this is missing
an href, it gets highjacked instead of delegating to React.

I've demoed this working to Marc over slack.