Fix issue where default values wouldn't be saved into the formData until edited
You can test this locally here: http://localhost:9966/?selected=swagger-files%2Ftypes.json
Search for "default value" and see that it is already in the code sample.

This issue luckily got fixed upstream:
https://github.com/mozilla-services/react-jsonschema-form/pull/1034

So to bring that in I needed to update my fork of the module.

I would love to get off of my fork and get back onto the main release,
but there's some outstanding work on my PR which means we can't right now:
https://github.com/mozilla-services/react-jsonschema-form/pull/954