Fix problem where nested objects would appear empty
On production right now if there's a nested object in a form, it always get
in the code sample even if the form values are not set:

```
--data '{"category":{}}'
```

This PR recursively removes empty objects from the form values so they
do not get set when making the request.

This problem was slightly exacerbated by #179 which auto applied default
values from the form to the samples.