fix jsinspect issue
Apparently my additional wrapper tag pushed `jsinspect` over it's 50% threshold so it failed the build... I refactor the `ResponseExample` CSS selectors to avoid this, since this was really a small change and this markup was already a bit specious! I do agree that this type of duplication _definitely_ smells. (A core set of ReadMe components might help ease some of this friction! https://github.com/readmeio/styleguidist-demo, not to plug my own pet project or anything...)