hotfix: application/x-www-form-urlencoded bug (#818)
* fix: issue where application/x-www-form-urlencoded payloads weren't being handled properly (#807)

* test: added failing test case for formData issue

I added two test cases: one for json and one for formData. The json
one passes fine! This code in oas-to-har is working as expected:
https://github.com/readmeio/api-explorer/blob/59a5d060997321936ae356f2ba3f36e34c81c520/packages/oas-to-har/src/index.js#L198

I narrowed down the issue to here:
https://github.com/Kong/httpsnippet/blob/633b825364ab050f02556cfa837148a1e136fd69/src/index.js#L126-L128
httpsnippet is setting it from the value we set back to ""

I'm not sure if this is something that's changed recently in the way
we're generating hars? This part of httpsnippet hasn't been updated
in 5 years. Gunna defer to Jon for this.

* fix: correctly preparing `x-www-form-urlencoded` forms as params

* fix: oas-to-har will now output valid HAR definitions

* chore(deps): upgrading fetch-har to 3.0.0

* docs: adding an example for our handling of formData

Co-authored-by: Jon Ursenbach <jon@ursenba.ch>

* build: updating dist files

* v6.12.1

Co-authored-by: domharrington <domharrington@users.noreply.github.com>