refactor: move operation extension logic out of Doc Component and into `oas-extension` package.
originally, we had the logic for using the operation level extension or the oas level extension in the Doc component. this caused the Parameter component tests to fail so we moved the logic to the `oas-extension` package. it now also sets a default if there isn't an extension on the oas and the operation level.